+++ src/main/java/com/codeup/adlister/dao/MySQLUsersDao.java
@@ -57,7 +57,7 @@ public class MySQLUsersDao implements Users {
             return null;
         }
         return new User(
-            rs.getLong("id"),
+            rs.getInt("id"),
             rs.getString("username"),
             rs.getString("email"),
             rs.getString("password")
